+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ## [0.10.1] 2019-11-14
+### Added
+- set-hooks: add support for pre/post start/stop hooks. Script are executed in the host environment (#61)
+
 ### Fixed
 - create: permission of /tmp in single type are wrong (#72)